fix cppcheck again

+++ src/adjmat/DFSOutput.cpp
@@ -108,7 +108,7 @@ void DFSOutput::update(){
           if( atoms[i] ) nats++;
       }
 
-     ofile.printf("CLUSTERING RESULTS AT TIME %f : NUMBER OF ATOMS IN %u TH LARGEST CLUSTER EQUALS %u \n",getTime(),clustr,myatoms.size() );
+     ofile.printf("CLUSTERING RESULTS AT TIME %f : NUMBER OF ATOMS IN %u TH LARGEST CLUSTER EQUALS %u \n",getTime(),clustr,static_cast<unsigned>(myatoms.size()));
      ofile.printf("INDICES OF ATOMS : ");
      for(unsigned i=0;i<myatoms.size();++i) ofile.printf("%d ",(myclusters->getAbsoluteIndexOfCentralAtom(myatoms[i])).index());
      ofile.printf("\n");
@@ -119,7 +119,7 @@ void DFSOutput::update(){
      }
      ofile.printf("\n");
   } else {
-     ofile.printf("CLUSTERING RESULTS AT TIME %f : NUMBER OF ATOMS IN %u TH LARGEST CLUSTER EQUALS %u \n",getTime(),clustr,myatoms.size() );
+     ofile.printf("CLUSTERING RESULTS AT TIME %f : NUMBER OF ATOMS IN %u TH LARGEST CLUSTER EQUALS %u \n",getTime(),clustr,static_cast<unsigned>(myatoms.size()) );
      ofile.printf("INDICES OF ATOMS : ");
      for(unsigned i=0;i<myatoms.size();++i) ofile.printf("%d ",(myclusters->getAbsoluteIndexOfCentralAtom(myatoms[i])).index());
      ofile.printf("\n");
